+++ src/main/java/si/matjazcerkvenik/alertmonitor/webhook/WebhookBean.java
@@ -254,8 +254,6 @@ public class WebhookBean {
                 }
             }
         }
-
-
 	}
 
 	private int getNumberOfSelectedTags() {